Improve dotfile subfolder structure and install procedure

+++ install.sh
@@ -5,21 +5,26 @@
 
 dir=~/dotfiles
 olddir=~/dotfiles_backup		#old dotfiles backup 
-files="vimrc vim zshrc tmux.conf tmux.reset.conf"
+file_paths="vim zsh/zshrc tmux/tmux.conf tmux/tmux.reset.conf"
 
 # Create dotfiles_old in homedir
-echo -n "Creating $olddir for backup of any existing dotfiles in home directory... "
-mkdir -p $olddir
-echo "Done"
+if [ ! -d "$olddir" ]; then
+	echo -n "Creating $olddir for backup of any existing dotfiles in home directory... "
+	mkdir -p $olddir
+	echo "Done"
+fi
 
 # Change to dotfiles directory
 echo -n "Changing to the $dir directory... "
 echo "Done"
 
 # Move any existing dotfiles to the backup directory, then creates symlinks to new files in home directory
-for file in $files; do
-	echo "Moving existing dotfile $file from ~ to $olddir."
-	mv ~/.$file $olddir
-	echo "Creating symlink to $file in home directory."
-	ln -s $dir/$file ~/.$file
+for path in $file_paths; do
+	file=$(basename $path)
+	if [ -e "$file" ]; then
+		echo "Moving existing dotfile $file from ~ to $olddir."
+		mv ~/.$file $olddir
+	fi
+	echo "Creating symlink to $path in home directory."
+	ln -sfn "$dir/$path" ~/.$file
 done
